Pull up MonkeyRunnerExtension class
